requirements.txt
langchain
ipykernel
langchain_community
langchain_huggingface
langchain_openai 
langchain_core
python-dotenv
streamlit

langserve
fastapi
uvicorn
sse_starlette
bs4
pypdf
chromadb
faiss-cpu
groq
cassio
beautifulsoup4
langchain-groq
wikipedia
arxiv
langchainhub
sentence_transformers
PyPDF2
langchain-objectbox
